Added support for ignoring false positives for crawler
