+++ package.json
@@ -143,7 +143,7 @@
     "stylelint": "stylelint --quiet --allow-empty-input 'packages/**/*.css' 'docs/**/*.css'",
     "test": "node --max_old_space_size=8192 node_modules/@ckeditor/ckeditor5-dev-tests/bin/test.js",
     "manual": "node --max_old_space_size=8192 node_modules/@ckeditor/ckeditor5-dev-tests/bin/test-manual.js",
-    "manual:verify": "node ./scripts/web-crawler.js --manual",
+    "manual:verify": "node ./scripts/web-crawler/index.js --manual",
     "bootstrap": "yarn install",
     "clean": "node ./scripts/remove-lock-files.js && rm -rf ./node_modules && rm -rf ./external/*/node_modules",
     "reset": "rm -rf ./node_modules && yarn run bootstrap",
@@ -154,7 +154,7 @@
     "docs:build-and-publish-nightly": "node ./scripts/docs/build-and-publish-nightly.js",
     "docs:content-styles": "node ./scripts/docs/build-content-styles.js",
     "docs:serve": "http-server ./build/docs/",
-    "docs:verify": "node ./scripts/web-crawler.js --docs",
+    "docs:verify": "node ./scripts/web-crawler/index.js --docs",
     "translations:collect": "ckeditor5-dev-env-translations collect",
     "translations:download": "ckeditor5-dev-env-translations download",
     "translations:upload": "ckeditor5-dev-env-translations upload",
