+++ packages/ckeditor5-core/tests/manual/version-collision.html
@@ -1,6 +1,9 @@
 <head>
 	<!-- "script-src 'unsafe-inline'" is required if we want to execute the inline <script> -->
 	<meta http-equiv="Content-Security-Policy" content="default-src 'none'; connect-src 'self' https://cksource.com http://*.cke-cs.com; script-src 'self' 'unsafe-inline' https://cksource.com; img-src * data:; style-src 'self' 'unsafe-inline'; frame-src *">
+	<meta name="x-cke-crawler-ignore-patterns" content='{
+		"uncaught-exception": "CKEditorError: ckeditor-duplicated-modules"
+	}'>
 </head>
 <script>
 	window.CKEDITOR_VERSION = 'the.colliding.version';
