+++ packages/ckeditor5-engine/tests/view/manual/noselection-iframe.html
@@ -1,1 +1,13 @@
-<iframe id="iframe" src="immutable.html"></iframe>
+<head>
+    <!--
+		For a totally unknown reason, loading a large script inside an iframe is aborted in Puppeteer,
+        so this request has been added to exclusions for the crawler ¯\_(ツ)_/¯.
+	-->
+	<meta name="x-cke-crawler-ignore-patterns" content='{
+		"request-failure": "immutable.js"
+	}'>
+</head>
+
+<body>
+    <iframe id="iframe" src="immutable.html"></iframe>
+</body>
