+++ packages/ckeditor5-utils/tests/manual/duplicated-import/duplicated-import.html
@@ -1,1 +1,7 @@
+<head>
+	<meta name="x-cke-crawler-ignore-patterns" content='{
+		"uncaught-exception": "CKEditorError: ckeditor-duplicated-modules"
+	}'>
+</head>
+
 <div id="editor"></div>
