+++ tests/manual/performance/editorinit.html
@@ -1,3 +1,9 @@
+<head>
+	<meta name="x-cke-crawler-ignore-patterns" content='{
+		"request-failure": "example.test"
+	}'>
+</head>
+
 <style>
 	.editor {
 		display: none;
